\encoding{UTF-8}
\name{bezerra2009}
\alias{bezerra2009}
\docType{data}

\title{Individuals observed in a flower-visitation network of oil-collecting bees in a Brazilian steppe.}

\description{

Observation of 38 individual plants from 13 oil-flower species of the family Malphighiaceae. Flower visitors were collected. Only legitimate visitors were considered. Numbers in cells refer to the amount of visitors of each bee species collected in each flower species.

The species interaction matrix describes the number of bee visits to 138 individual plants in natural clumps of 13 Malpighiaceae species during the flowering peak of each species. The number of bee visits  to flowers was registered over four consecutive days, from 5.00 to 17.00 with a total of 1392 h of observations. 

Location: Parque Nacional do Catimbau, Brazil (8째24'00'' - 37째 36'35''S and 3째 09'30'' - 37째 14'40''W)

Biome: Caatinga (Brazilian steppe)

The paper itself contrasted a network of Malpighiaceae oil-flowers and associated oil-collecting bees from a Brazilian steppe (``caatinga'') to whole pollination networks from all over the world available in the Interaction Web Database. The caatinga network had a perfectly balanced proportion of plants and animals (13 x 13) and was more nested and less modular than all of the 22 whole pollination networks studied. The authors concluded that the oil-flower subweb is more cohesive and resilient than whole pollination networks, reinforcing the hypothesis that each ecological service is in fact a mosaic of different subservices with a hierarchical structure (``webs within webs'').      

Take from the NCEAS interaction web database (\url{http://www.nceas.ucsb.edu/interactionweb}).

}

\usage{data(bezerra2009)}

\references{
Bezerra, E.L.S., Machado, I.C.S. and Mello, M.A.R. 2009. Pollination networks of oil-flowers: a tiny world within the smallest of all worlds. \emph{Journal of Animal Ecology} \bold{78}, 1096--1101.
}

\examples{
data(bezerra2009)
}
\keyword{datasets}
