% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-bipartite-graph-nie.R
\name{learn_bipartite_graph_nie}
\alias{learn_bipartite_graph_nie}
\title{Laplacian matrix of a k-component bipartite graph via Nie's method

Computes the Laplacian matrix of a bipartite graph on the basis of an observed similarity matrix.}
\usage{
learn_bipartite_graph_nie(
  S,
  r,
  q,
  k,
  learning_rate = 1e-04,
  eta = 1,
  maxiter = 1000,
  reltol = 1e-06,
  verbose = TRUE,
  record_objective = FALSE
)
}
\arguments{
\item{S}{a p x p similarity matrix, where p is the number of nodes in the graph.}

\item{r}{number of nodes in the objects set.}

\item{q}{number of nodes in the classes set.}

\item{k}{number of components of the graph.}

\item{learning_rate}{gradient descent parameter.}

\item{eta}{rank constraint hyperparameter.}

\item{maxiter}{maximum number of iterations.}

\item{reltol}{relative tolerance as a convergence criteria.}

\item{verbose}{whether or not to show a progress bar during the iterations.}

\item{record_objective}{whether or not to record the objective function value during iterations.}
}
\value{
A list containing possibly the following elements:
\item{\code{laplacian}}{estimated Laplacian matrix}
\item{\code{adjacency}}{estimated adjacency matrix}
\item{\code{B}}{estimated graph weights matrix}
\item{\code{maxiter}}{number of iterations taken to reach convergence}
\item{\code{convergence}}{boolean flag to indicate whether or not the optimization converged}
\item{\code{obj_fun}}{objective function value per iteration}
}
\description{
Laplacian matrix of a k-component bipartite graph via Nie's method

Computes the Laplacian matrix of a bipartite graph on the basis of an observed similarity matrix.
}
\references{
Feiping Nie, Xiaoqian Wang, Cheng Deng, Heng Huang. Learning A Structured Optimal Bipartite Graph for Co-Clustering.
            Advances in Neural Information Processing Systems (NIPS 2017)
}
