% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-gaussian-bipartite-pgd.R
\name{learn_connected_bipartite_graph_pgd}
\alias{learn_connected_bipartite_graph_pgd}
\title{Laplacian matrix of a connected bipartite graph with Gaussian data

Computes the Laplacian matrix of a bipartite graph on the basis of an observed data matrix.}
\usage{
learn_connected_bipartite_graph_pgd(
  S,
  r,
  q,
  init = "naive",
  learning_rate = 1e-04,
  maxiter = 1000,
  reltol = 1e-05,
  verbose = TRUE,
  record_objective = FALSE,
  backtrack = TRUE
)
}
\arguments{
\item{S}{a p x p covariance matrix, where p is the number of nodes in the graph.}

\item{r}{number of nodes in the objects set.}

\item{q}{number of nodes in the classes set.}

\item{init}{string denoting how to compute the initial graph.}

\item{learning_rate}{gradient descent parameter.}

\item{maxiter}{maximum number of iterations.}

\item{reltol}{relative tolerance as a convergence criteria.}

\item{verbose}{whether or not to show a progress bar during the iterations.}

\item{record_objective}{whether or not to record the objective function value during iterations.}

\item{backtrack}{whether or not to optimize the learning rate via backtracking.}
}
\description{
Laplacian matrix of a connected bipartite graph with Gaussian data

Computes the Laplacian matrix of a bipartite graph on the basis of an observed data matrix.
}
