% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-heavy-tail-bipartite-pgd.R
\name{learn_heavy_tail_bipartite_graph_pgd}
\alias{learn_heavy_tail_bipartite_graph_pgd}
\title{Laplacian matrix of a connected bipartite graph with heavy-tailed data

Computes the Laplacian matrix of a bipartite graph on the basis of an observed data matrix
whose distribution is assumed to be Student-t.}
\usage{
learn_heavy_tail_bipartite_graph_pgd(
  X,
  r,
  q,
  nu = 2.001,
  learning_rate = 1e-04,
  maxiter = 1000,
  reltol = 1e-05,
  init = "default",
  verbose = TRUE,
  record_objective = FALSE,
  backtrack = TRUE
)
}
\arguments{
\item{X}{a n x p data matrix, where p is the number of nodes in the graph and n is the number of observations.}

\item{r}{number of nodes in the objects set.}

\item{q}{number of nodes in the classes set.}

\item{nu}{degrees of freedom of the Student-t distribution.}

\item{learning_rate}{gradient descent parameter.}

\item{maxiter}{maximum number of iterations.}

\item{reltol}{relative tolerance as a convergence criteria.}

\item{init}{string denoting how to compute the initial graph or a r x q matrix with initial graph weights.}

\item{verbose}{whether or not to show a progress bar during the iterations.}

\item{record_objective}{whether or not to record the objective function value during iterations.}

\item{backtrack}{whether or not to optimize the learning rate via backtracking.}
}
\description{
Laplacian matrix of a connected bipartite graph with heavy-tailed data

Computes the Laplacian matrix of a bipartite graph on the basis of an observed data matrix
whose distribution is assumed to be Student-t.
}
