\name{memmott1999}
\alias{memmott1999}
\docType{data}
\title{ Flower visitation network from a meadow near Bristol, UK}
\description{
This study was conducted in a 150 x 250 m meadow plot in the vicinity of Bristol, U.K. in July 1997. The objective was to describe the plant-flower visitor interaction web of this area, taking into account species abundances and their frequency of interaction. Twenty five plant species were studied, and 79 flower visitor species were recorded visiting them.
}
\usage{data(memmott1999)}
\details{
The author recorded her data by counting the number of visits of each flower visitor species to each plant species, and by independently measuring the abundance of plant and animal taxa. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.
}
\source{
NCEAS
}
\references{
Memmott, J. 1999. The structure of a plant-pollinator food web. \emph{Ecology Letters}\bold{2}, 276--280.
}
\examples{
data(memmott1999)
## maybe str(memmott1999) ; plot(memmott1999) ...
}
\keyword{datasets}
