\name{motten1982}
\alias{motten1982}
\docType{data}
\title{A spring flower visitation network from North Carolina, USA}
\description{
This is a study of the interactions between insects visitors and spring wildflowers in piedmont North Carolina. Spring flowering, entomophilous herbs, shrubs, and understory trees were included in the study.

}
\usage{data(motten1982)}
\details{
The author recorded his data by counting the number of visits of each flower visitor species to each plant species. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.
}
\references{
Motten, A.F. (1982) Pollination Ecology of the Spring Wildflower Community in the Deciduous Forests of Piedmont North Carolina. Doctoral Dissertation thesis, Duke University, Durham, North Carolina, USA.

Motten, A.F. (1986) Pollination ecology of the spring wildflower community of a temperate deciduous forest. \emph{Ecological Monographs} \bold{56}, 21--42.

}
\examples{
data(motten1982)
## maybe str(motten1982) ; plot(motten1982) ...
}
\keyword{datasets}
