\name{olesen2002aigrettes}
\alias{olesen2002aigrettes}
\docType{data}
\title{ A flower visitation network from the Azores}
\description{
The structure of pollination networks were investigated for two oceanic islands, the Azorean Flores and the Mauritian Ile aux Aigrettes.
}
\usage{data(olesen2002aigrettes)}

\details{
The authors recorded their data by counting the number of visits of each flower visitor species to each plant species. Data are presented as an interaction frequency matrix, in which cells with positive integers indicate the frequency of interaction between a pair of species, and cells with zeros indicate no interaction.
}

\references{
Olesen, J. M., L. I. Eskildsen, and S. Venkatasamy. 2002. Invasion of pollination networks on oceanic islands: importance of invader complexes and endemic super generalists. \emph{Diversity and Distributions} \bold{8}, 181--192.
}
\examples{
data(olesen2002aigrettes)
## maybe str(olesen2002aigrettes) ; plot(olesen2002aigrettes) ...
}
\keyword{datasets}
