\encoding{UTF-8}
\name{use_deprecated_plotweb}
\alias{use_deprecated_plotweb}

\title{Switch between the current and the old (deprecated) plotweb function.}
\description{
Calling this function let's you use the old (deprecated) plotweb function, 
without explicitly calling `plotweb_deprecated`.
}
\usage{
use_deprecated_plotweb(deprecated = TRUE)
}

\arguments{
  \item{deprecated}{If `TRUE`, `plotweb` will become `plotweb_deprecated` }
}



