\name{vazcer}
\alias{vazcer}
\docType{data}
\title{A pollination network.}
\description{
 [See documentation for Safariland.]
}
\usage{data(vazcer)}
\format{
  The format is:
 num [1:9, 1:33] 59 0 110 0 0 0 0 0 0 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:9] "X2" "X1" "X13" "X4" ...
  ..$ : chr [1:33] "78" "65" "94" "32" ...
}
\examples{
data(vazcer)
## maybe str(vazcer) ; plot(vazcer) ...
}
\keyword{datasets}
