\name{vazquec}
\alias{vazquec}
\docType{data}
\title{A pollination network.}
\description{
 [See documentation for Safariland.]
}
\usage{data(vazquec)}
\format{
  The format is:
 num [1:8, 1:27] 31 0 34 150 11 66 2 4 0 63 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:8] "X1" "X13" "X4" "X15" ...
  ..$ : chr [1:27] "65" "94" "32" "66" ...
}
\examples{
data(vazquec)
## maybe str(vazquec) ; plot(vazquec) ...
}
\keyword{datasets}
