\name{vazquenc}
\alias{vazquenc}
\docType{data}
\title{A pollination network.}
\description{
 [See documentation for Safariland.]
}
\usage{data(vazquenc)}
\format{
  The format is:
 num [1:7, 1:24] 89 0 404 0 0 0 0 0 147 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:7] "X2" "X1" "X13" "X15" ...
  ..$ : chr [1:24] "78" "65" "32" "66" ...
}
\details{

}
\source{

}
\references{

}
\examples{
data(vazquenc)
## maybe str(vazquenc) ; plot(vazquenc) ...
}
\keyword{datasets}
